Use unix time stamps instead of Instant
kotlinx.datetime under the hood calls some Java APIs not available on Android API 25 and lower. This forces clients using this library to use desugaring.
We don't wan't that. And since time calculations are not very crucial in this project, we just use seconds since epoch instead, which can just be a Long.

+++ auth/src/main/kotlin/com/tidal/sdk/auth/model/RefreshResponse.kt
@@ -9,7 +9,7 @@ internal data class RefreshResponse(
     val accessToken: String,
     val clientName: String? = null,
     @JsonNames("expires_in")
-    val expiresIn: Int,
+    val expiresIn: Long,
     @JsonNames("token_type")
     val tokenType: String,
     @JsonNames("scope")
